+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # JRuby openHAB Scripting Change Log
 
+## [5.0.1](https://github.com/openhab/openhab-jruby/compare/v5.0.0...main)
+
+### Features
+
+- {OpenHAB::Core::Rules::Registry#scenes rules.scenes} and {OpenHAB::Core::Rules::Registry#scripts rules.scripts}
+  are available as convenient shortcuts to filter for rules that are scenes or scripts.
+
 ## [5.0.0](https://github.com/openhab/openhab-jruby/compare/4.45.2...v5.0.0)
 
 5.0 is the first release as the officially supported gem of the openHAB organization.